fixed the contructor injection for classlynkapplicaiton
