+++ src/main/java/ai/classlynk/entity/OptimizationParameter.java
@@ -1,0 +1,5 @@
+package ai.classlynk.entity;
+
+public interface OptimizationParameter {
+
+}
